@@ -5,6 +5,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using MyMoods.Hangfire.Settings;
 using MyMoods.Services.Injection;
+using Serilog;
 using System.IO;
 using System.Reflection;
 
@@ -14,7 +15,7 @@ namespace MyMoods
     {
         private readonly IConfigurationRoot _configuration;
 
-        public Startup(IHostingEnvironment env)
+        public Startup(IWebHostEnvironment env)
         {
             var builder = new ConfigurationBuilder()
                 .SetBasePath(Path.GetDirectoryName(Assembly.GetEntryAssembly().Location))
@@ -30,9 +31,15 @@ namespace MyMoods
             DefaultInjection.Inject(services, _configuration);
         }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env, ILoggerFactory loggerFactory)
         {
-            DefaultLogger.Use(env, loggerFactory);
+            var logger = new LoggerConfiguration()
+              .MinimumLevel.Information()
+              .WriteTo.RollingFile(Path.Combine(env.ContentRootPath, @"logs/{Date}.log"))
+              .CreateLogger();
+
+            loggerFactory.AddSerilog(logger);
+
             HangfireConfiguration.Configure(app, _configuration);
         }
     }
