@@ -7,6 +7,7 @@ using MyMoods.Services.Injection;
 using MyMoods.Settings;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
+using Serilog;
 using System.IO;
 using System.Reflection;
 
@@ -16,7 +17,7 @@ namespace MyMoods
     {
         private readonly IConfigurationRoot _configuration;
 
-        public Startup(IHostingEnvironment env)
+        public Startup(IWebHostEnvironment env)
         {
             var builder = new ConfigurationBuilder()
                 .SetBasePath(Path.GetDirectoryName(Assembly.GetEntryAssembly().Location))
@@ -31,17 +32,25 @@ namespace MyMoods
         {
             DefaultInjection.Inject(services, _configuration);
 
-            services.AddMvc();
-            services.AddMvcCore().AddJsonFormatters(x => ConfigureJson(x));
+            services.AddControllers()
+                .AddNewtonsoftJson(options =>
+                {
+                    options.SerializerSettings.Converters.Add(new StringEnumConverter());
+                    options.SerializerSettings.ReferenceLoopHandling = ReferenceLoopHandling.Ignore;
+                    options.SerializerSettings.PreserveReferencesHandling = PreserveReferencesHandling.None;
+                });
         }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env, ILoggerFactory loggerFactory)
         {
-            DefaultLogger.Use(env, loggerFactory);
+            var logger = new LoggerConfiguration()
+              .MinimumLevel.Information()
+              .WriteTo.RollingFile(Path.Combine(env.ContentRootPath, @"logs/{Date}.log"))
+              .CreateLogger();
+
+            loggerFactory.AddSerilog(logger);
 
             app.UseMiddleware<AnalyticsAuthorizationMiddleware>();
-
-            app.UseMvc();
             app.UseDefaultFiles();
             app.UseStaticFiles();
         }
